Reverse calculation order in getTopAnalyzers
The order of operands in the difference calculation method in getTopAnalyzers function has been reversed. This change will now subtract the 'startCapital' from the 'money' instead of vice versa. This modification ensures the correct value is used for further percentage difference calculations and the data is accurately sorted in descending order.