Add create time to AccountResponse and update UI
Added the `createTime` property to the AccountResponse data class in AccountController.kt and AccountResponse.kt. A new 'Account Creation Date' section was added in AccountPage.tsx to display this new information, with corresponding style updates in new 'AccountInfo.css'. The Account model in 'Account.tsx' was also updated to include 'createTime'.