Add API token creation functionality
Created an end-to-end functionality for creating and adding API tokens. A new create API token request model was added on the server-side. Added an API endpoint to handle the server-side token creation logic. On the client-side, a dialog was created for user interactions to add API tokens. To trigger this dialog, a button was added into the user account page. The CSS was revised to support these UI updates.