Add retry mechanism to HTTP requests
The code has been updated to add a retry mechanism for HTTP requests in the event of a HttpRequestTimeoutException. This mechanism can retry up to 2 times by default, and the number can be customized on each HTTP request function call. The changes affect multiple functions in the ByBitPrivateHttpClient and TradeManager classes.