Improve logging and change log level to INFO
This update brings changes to the logging mechanism in the TradeManager class. Log levels for  "local" spring profiles were also set to INFO in logback-spring.xml files in trade-manager, api-service, analyzer-service, and event-listener directories. The primary improvement was wrapped the logger with a method which automatically adds and clears the "manager-id" Mapped Diagnostic Context (MDC) before and after the logging action.