Update Logstash condition in docker-compose
Changed the Logstash dependency condition setting from 'service_started' to 'service_healthy' in the docker-compose.prod.yml file. This change will ensure that dependent services only start after Logstash is not just started, but is also in a healthy running state. It aims to improve the stability of the application environment.