Implement real-time analyzer and manager data through websockets
Introduced websocket communication to provide real-time data for analyzer and manager services. Added a web socket module on both UI and backend side to allow continuous data exchange. Also, slight adjustments were made to other sections to accommodate these changes like updating TopicPartitionOffset in PriceTickerListener and changing dependencies in build.gradle.kts. This enhancement provides active updates on the state of the analyzer and manager, increasing the usability and interactivity of the service.