Optimize analyzer class and enhance code readability
The code in the analyzer class was refactored to improve efficiency and security. The snapshots variable was transformed into a list before processing to prevent repeated conversions thus enhancing the performance. The code readability was enhanced by simplifying the 'createTree' method in the 'FindRangeUtils' class and reducing the visibility of the 'stabilityCoef' variable to private, which also enhances the security of data access.