Refactor method name and clear position post-close
Renamed the method 'requestMostProfitableAnalyzer' to 'requestAnalyzer' to better reflect its role and not limit its scope to most profitable analyzers only. Also added a code line in the Manager class to nullify the position object once closed, for cleaner state management and to prevent any unexpected behaviour because of stale position references.