Add sign to percentage changes in AnalyzerContent
Functionality to prepend '+' and '-' signs according to positive or negative percentage changes has been added. This is achieved by introducing a new function 'getSignByValue', which provides the required signs based on the value. It is applied to 'pnl1', 'pnl12', and 'pnl24' values in the AnalyzerContent file, aiding the better understanding of data trends.