Adjust stabilityCoef filter and enhance trade-manager functionality
This update enhances AnalyzerService by reducing the stabilityCoef filter from 4 to 2, thereby considering more analyzers for processing. In the trade-manager, the setInitialCumRealizedPnL() function removal, addition of setDiapasonConfigs(), altered bybitService.closePosition() parameters, and a change in the isPositionEmpty() function altogether improve order handling and strategy execution. The strategy-lib also sees the addition of 'setDiapasonConfigs()' for configuring the price range.