Update stabilityCoef filter and improve trade-manager functions
In this update, the stabilityCoef filter in AnalyzerService got adjusted from 4 to 2 to broaden the scope of analyzers for processing. A multitude of modifications were carried out in the trade-manager, including the removal of setInitialCumRealizedPnL() function, addition of setDiapasonConfigs(), alteration in parameters of bybitService.closePosition(), and a change in the isPositionEmpty() function to improve order management and strategy enforcement. The 'setDiapasonConfigs()' function was also introduced to the strategy-lib for better price range configurations.