[CBA-79] Need to add auto list loading to append analyzers list
Add infinite scrolling to analyzer table

A feature to load more data as the user scrolls down in the analyzer list page has been added. This implementation involves usage of 'react-infinite-scroller' library, due to which both the package and its type definitions have been added to the dependencies. The UI states and styles have been changed to accommodate this feature.