Update ContactPage UI and enhance AnalyzerService functionality
In the ContactPage UI, full-logo SVG import and the presentation of Discord and Contact Us links have been updated. In the backend, AnalyzerService has been enhanced to consider the stability of active analyzers by referring to the analyzerStabilityRepository. Moreover, unnecessary conditions in AnalyzerContent that were guarding calls to deleteAnalyzers and resetAnalyzers functions have been removed, simplifying the code.