Add money amount condition to strategy functions
The code has been updated to include a condition that checks if the `money` variable is greater than zero before executing the body of strategy functions in the `CandleTailStrategyRunner` class. This change helps in preventing needless calculations and function calls when there is no money, thus enhancing the program's performance. Also, when money equals zero, it is now set to zero directly to avoid any potential floating-point precision issues.