Refactor strategy runners for clearer logic and cleaner code
The commit includes changes that make the strategy runners' code clearer and more efficient. The main change is ensuring that case matching occurs in a cleaner and more readable fashion. Another improvement is the streamlining of the logic in the CandleTailStrategyRunner. Predicting the money to use has been relocated so that it can be computed less redundantly, resulting in more efficient code. Other code quality improvements in spaces and line breaks are also part of the commit.