Update AnalyzerService and CandleTailStrategyRunner logic
The AnalyzerService.kt file was updated with a note to add filtering by folder in the future. The CandleTailStrategyRunner.kt file has been adjusted to add a utility for calculating percentage difference, as well as a note for a future check to ensure that qty and price are greater than the min qty and min price. The moneyToUse logic has been tweaked in the CandleTailStrategyRunner to limit the maximum investment to 70% of funds.