Update condition and callback invocation in ByBitWebSocketClient
The condition for invoking callbacks in ByBitWebSocketClient.kt has been changed from checking if both callbacks are not null to checking if either one is not null. This allows individual callbacks to be invoked independently. Also, the DEBUG level logging for the ByBit integration client has been removed from application-prod.yml to reduce log verbosity.