Disable analyzer creation check for users
The check that previously ensured a user could create analyzers based on their account ID and the size of ID list has been commented out. This was done in both the validateAnalyzersExistByIdAndAccountId and setAnalyzersActiveStatus functions in AnalyzerService.kt. From this point forward, users are presumed to have the default capability to create analyzers.